chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cairosvg>=2.8.2",
    "chess>=1.11.2",
    "picamera2>=0.3.30",
    "picamzero>=0.0.2",
    "python-dotenv>=1.1.1",
    "slack-bolt>=1.24.0",
]
